(ss_table): quick speedup and rename
This commit fixes a gotcha in which the memtable size is incremented
even for keys that already exist in it. By fixing this bug, the
pre-existing memtable nodes can simple be updated in place and the
method can exit. This is a huge speedup for the benchmarks.

Some methods are also renamed to drop coments and improve code
redability.